fixup: Enable GASNET_VIS_{AMPIPE,REMOTECONTIG} by default
02556217979a30340b75c9156e03947bc15dc424 changed the default
value of GASNET_DISABLE_MUNMAP to 1 for ibv-conduit on all platforms,
but this was too aggressive since this hack doesn't work well on
ILP32, where pmalloc easily exhausts available sbrk space and then
ignores the setting. Scale back to only enabling this by default on
LP64, where it seems to work well.

This fixes a regression observed on ibv-32/testcollperf-par.