NP-AM: corrections to usage of max payload queries
This commit corrects two types of incorrect argument usage in Prepare
code (reference, GNI and template) when making max payload queries.

First, this commit adds logic to strip the (documented as ignored)
LEAST bits from 'flags' upon entry to Prepare.  This is needed to at
least ensure they are not passed to a max payload query, where they
would result in obtaining the wrong 'limit' value.  By masking them
out early we ensure they are not mistakenly passed *anywhere* that
isn't expecting to handle them.

Second, logic is added to ensure that (lc_opt == pointer-to-event) is
mapped to NULL in calls to the error-checking max payload queries.
This is sensible in the (non-performant) conduit-independent reference
code where no assumption can safely be made about the conduit.
However, the case of the GNI-specific code we (continue to) use the
internal call (which performs no error-checking, but is known to ignore
the lc_opt) with the lc_opt argument as passed to the Prepare.