coll: replace gasnet_coll_handle_t w/ gex_Event_t
This commit makes numerous changes to bring the existing GASNet-1
collectives implementation into the event-based GEX infrastructure.

In GASNet-1, there is a distinct handle type for collective ops, with
distinct sync ops: gasnet_coll_{try,wait}_sync{,_all,_some}().  After
this commit gasnet_coll_handle_t is an alias (typedef) for gex_Event_t
and the family of sync calls are thin wrappers around the gasnete_*
sync calls (retained to preserve the historical tracing output).
Internal callers use the gasnet_* sync calls directly.

Historically in-flight collective ops have been progressed *only* via
the coll-specific sync calls.  With their removal some other means is
needed to ensure collectives progress.  So, this commit introduces a
boolean progress function.  Boolean is used, rather than counted,
because the collectives infrastructure maintains a linked-list of
in-flight ops and observing the empty/non-empty transitions (either
direction) is simple.

Independent of the switch to gex_Event_t for representing collective
ops, their use for point-to-point (RMA and VIS) has eliminated the
need to ensure that only the initiating thread will test stored
events.  Therefore, the thread-specific handles table is replaced with
a global table of in-flight events.  Polling this table was already
serialized prior to this change.  Additionally, the array-of-structs
has been replaced with a struct-of-arrays to allow proper use of
test-some (reducing overhead).

With the removal of thread-specific handle requirements, there is no
longer a need to track an 'owner' of a collective operation, and so
the field and associated macros to set and query it are removed.

Note that no effort has been made perform an exhaustive update of all
uses of 'handle' in text or as identifiers.