ofi: Adjust inline directives to avoid warnings
gasnetc_ofi_am_recv_poll() is called in both "hot" and "cold" contexts:

* Hot: critical path in gasnet_AMPoll(), a good candidate for function inlining
* Cold: backpressure polling path

The calls on the "cold" path were leading to -Winline warnings
because the inlining heuristic (correctly) noticed the call path
was unlikely and thus not a profitable inlining candidate.

Split the old function (which had a PLEASE_INLINE directive) into
two new versions with force-INLINE and no inlining directives, and
adjust the call sites appropriately.

Correct a similar problem on fi_cq_readerr() which carries an
inline keyword in libfabric 1.6.0 and causes similar warnings
(because it's only called by the conduit on fatal/exit paths).

Measurements on dirac with gcc 4.8.5 and 7.3.0 w/sockets provider
show testam performance to be within noise before/after this change.