Remove GASNET[TI]_PLEASE_INLINE
As resolved in the 2018-2-22 meeting, our implementation of
GASNETT_PLEASE_INLINE never did what it advertised (expose an
optional inline decoration that is guaranteed to be warning-free).
We don't feel it's our place to expose a "wish-washy" inlining
directive. This is already available to C99 and C++ clients
via the "inline" standard keyword, and good optimizers will already
assume any function not explicitly decorated otherwise is a candidate
for potential inlining.

Advance the tools spec along with this deliberately breaking change.
We decided we want any potential clients of this removed functionality
to examine their call site and make an informed decision about whether
they meant to force inlining (in which case they should change to use
GASNETT_INLINE) or leave it wishy-washy (in which case they should
either delete the directive or possibly add an 'inline' keyword).