fixup: trace: add flag to AM REQ/REP tracing
This commit corrects ibv-conduit direct usage of the tracing macros
for AM injection to add the recently-added 'flags' argument.

Ibv-conduit's "system" AMs don't use the GASNETI_COMMON_AM*() macros,
but (at least for now) we include them in trace/stats.  Because they
are not using the GASNETI_COMMON_AM*() macros, they call the tracing
macros directly.  These calls were overlooked when the 'flags'
argument was added to tracing of AMs.