fixup: trace: add tracing of AM handler detail ...
This commit corrects the recently-added code to trace the client name for a
running handler to allow for the possibility of a NULL token_info.gex_ep,
as seen w/ conduit-internal AMs that at least gemini/aries may run prior to
gex_Client_Init().  The commit additionally allows for the possibility of a
non-NULL gex_ep having a NULL client.