Fix an automake misuse that broke distcheck-old
For some unknown reason, the top-level Makefile.am has always
included $(headers) in $(EXTRA_DIST), dating back to the first versions
of GASNet.  This is unnecessary (because headers are distributed by
default) and contrary to automake recommendations.

The recent addition to $(headers) of files in subdirs caused this to
trip a bug in automake 1.4 handling of $(EXTRA_DIST), so workaround it
by removing $(headers) from that variable where it never belonged in the
first place.

Other misc minor cleanups.