ibv: use HCA's maximum MTU by default
This commit changes the default InfiniBand MTU logic from using a
fixed 1K MTU, to using the largest MTU supported by the HCA(s).

Prior (by years) to this commit there was a concern that larger MTUs
were leading to increased latency for small RMA operations.  So, some
tests were performed before making this change.

On "old" HCAs (mlx4 generation on Dirac) no measurable difference was
seen among 1K, 2K and 4K MTUs for either latency or bandwidth reported
with testsmall and testlarge.

Meanwhile, mlx5 generation HCAs at JLSE and on SummitDev show nearly a
2x improvement in peak Get bandwidth compared to the old 1K default
(which is the motivation for this change).  No other changes to
latency or bandwidths were observed.

Testing with testam (with AMRDMA disabled) shows mixed results for AM
traffic.  The Short and Long results are unchanged, as are small
Mediums.  However for Mediums 1K and larger the switch to 4K produces
lowered bandwidths for Request-Reply ping-pong and increased
bandwidths for flood tests.  This is more pronounced with the mlx4 HCA
(approx 10% loss and 5% gain) than mlx5 (0.2% loss and 4% gain).