inline asm: IBM XL Compilers 12.0+ support GCC asm
This commit updates gasnet_asm.h to enable use of GCC-style inline ASM
on all IBM XL Compilers version 12.0.0 and newer.  Tests of 12.0.0 on
Linux, AIX and BG/Q all confirm support is present in both the C and
C++ compilers.

Ironically, this all but eliminates the need for the recent "unknown
compiler" work for timers and membars with respect to XLC.