tools: use "slow" membars for unknown compilers
This commit rewrites much of the preprocessor logic in gasnet_membar.h
to better handle compilers not known to us, whether at GASNet build
time or as an "unknown" compiler at client compile time.  In many
instances whitelists of compilers have been replaced with
GASNETI_HAVE_GCC_ASM (or similar) which are "feature tests".  This
should allow, for instance, support for Clang on platforms where we
have not yet explicitly enabled it.

In the case of a compiler for which we do not know how to generate
the necessary inline asm, we still #error when building the GASNet
libraries.  However, for unknown compilers at client compilation time
we now use "slow" membars (calls to functions in gasnet_tools.c).