Cleanup asm support for Sun/Oracle compilers
This commit resolves
  Bug 3748 - update/improve Solaris Sun C++ asm support

This commit makes the following changes to handling of inline assembly
with the Sun/Oracle compilers:

1. Adds a configure probe of Sun compilers to determine support for
`asm("mnemonic")`.  This has been supported in their C++ compilers on
Linux since that compiler was first released, and on Solaris/x86 since
at least "Sun Studio 12.0" (aka "Sun C++ 5.9", released in 2007).
However, if/when support was added for Solaris/SPARC is not known.
Keep in mind, this impacts the C++ compiler only, as the C compiler
has always supported `__asm(...)`.

2. Updates the logic in gasnet_asm.h to use the new configure probe
information and version knowledge, above.  The later should help to
increase use of inline assembly for Sun C++ used as an "unknown"
compiler.

3. Updates membar selection logic for Sun/Oracle compilers to use
GASNETI_ASM_AVAILABLE as set by gasnet_asm.h instead of trying to
duplicate that logic.  IN FACT, the logic in gasnet_membar.h has NOT
been kept in-sync and was consequently unconditionally using "slow"
(function call) membars with Sun C++ even with all the changes above.

4. Makes cosmetic changes to how gasnet_asm.h distinguishes between
the C and C++ cases (for readability/maintainability).