External CI: tweaks for --include=runnable
This commit makes changes to the harness.conf for the three "external"
CI suites: upcxx, chapel and legion.  The purpose is to ensure that
the builds of the respective runtime libraries and/or compilers will
match a harness filter option of "--include=runnable".  Otherwise, all
of the *actually* runnable tests will fail to build.