atomics: correct/explain 64-bit sync atomics usage
This commit corrects the preprocessor logic in gasnet_atomic_bits.h to
match gasnet_atomic_fwd.h with respect to use of 64-bit __sync
atomics.  Specifically we do *not* use them on ILP32 targets due to
(current) lack of knowledge of their signal safety on various
platform.

Additionally, this commit adds `#if 0`-guarded place-holders to help
any later use of __sync atomics which are known NOT to be signal-safe.