gasnet_asm.h: separation of concerns
This commit separates the preprocessor logic devoted to identifying
version-dependent behaviors in PGI from the preprocessor logic for
determining which inline ASM family is supported.  This is needed
to ensure the PGI behaviors are examined when GASNETI_HAVE_GCC_ASM
is set (e.g. by configure in an upcoming commit).