configure: sane defn for GASNETI_HAVE_*_GCC_ASM
This commit updates the logic in configure to always set
GASNETI_HAVE_{CC,CXX,MPI_CC}_GCC_ASM for compilers where we believe
such support exists.  Prior to this, only builds with IBM XL Compilers
would set these configure outputs!

The primary change is a simple duplication of the compiler whitelist
already present in gasnet_asm.h.  This change does NOT change which
compilers actually *use* GCC-style asm when building GASNet libraries.
However, this change corrects an error, introduced in PR208, in which
configure misidentified the atomics implementation of non-XLC
compilers on PPC as "Generic" (when, in fact, GCC, Clang and PGI all
support native atomics).  This would lead to failures of the
"mismatch" harness tests.

An additional change introduced in this commit is to split off
GASNET_CHECK_GCC_ASM() in configure from the XLC-specific portion
of GASNET_CHECK_ASM_SUPPORT().  That XLC-specific logic is still
the only caller, but this split will ease later addition of probes for
GCC-style asm support (such as anticipated for Sun C under bug 2941).