pshm/NP-AM: fix bug 3749
This commit resolves Bug 3749 (assertion failure in PSHM's NP-AM) by
removing the conditional sizing of the 'numbytes' field in the Medium
header for AMPSH.  This eliminates the overflow detected by the
assertion which failed in the bug report.

It is worth noting that due to padding for Medium payload alignment,
the size of the structure has not changed as a result of unconditional
use of a 4-byte field for the Medium payload length.

(cherry picked from commit edbf84b7782f3e62e847ccc613a9ba24be50ce1c)

Conflicts:
	gasnet_pshm.c