atomics: revised atomics compatibility logic
This commit replaces the recently-added logic designed to force
unknown compilers to use "slow" atomics.

The new logic expands coverage from just the case of unknown compilers
to also deal with possibility that CXX or MPI_CC have capabilities
which differ from those of CC.

In contrast to the previous logic, which ensured that unknown
compilers would always use SLOW or GENERIC, the new logic compares the
outputs of gasnet_atomic_fwd.h to those for the build CC (as recorded
at configure time).

If the gasnet_atomic_fwd.h outputs for both the 32- and 64-bit atomics
are "compatible" (see below) with those of CC then the new logic does
nothing to prevent the current compiler from traversing the giant

If the implementation selected for 32-bit atomics is not compatible,
then the logic will #define the identifiers needed to ensure the
ladder is traversed to reach the appropriate implementation which
could be GENERIC, OS or SLOW.

If the implementation selected for 64-bit atomics is not compatible,
then the logic will again #define the identifiers needed to ensure
that the appropriate implementation is built.  However, this is
complicated by the fact that the choice of 32-bit implementation has
control over how the ladder is traversed.

This complication is partially mitigated by the observation that there
are dependencies which limit the choice of 64-bit implementations once
the 32-bit one has been selected.  These observations have been
elevated to documented rules, and checks added to the end of
gasnet_atomic_fwd.h to #error if they are ever violated.

The remaining complication regarding generation of the correct 64-bit
implementation is addressed by adding `#if GASNETI_NOINLINE_ATOMIC64`
cases to the logic for each architecture (plus OS and SYNC) where we
already make LP64/ILP32 distinctions when generating 64-bit atomics.
That allows graceful handling of the case where the current compiler
should traverse the ladder to generate the same 32-bit implementation
as CC, but must use SLOW or GENERIC 64-bit atomics.  Compared to the
alternative of splitting the giant ladder into two distinct ones for
32-bit and 64-bit atomics, this was clearly a much smaller change to
make.

Re: "compatible"
As explained in more detail in the comments in the new logic, it was
found that we have been cautious enough regarding use of __sync
atomics that we have high confidence that they can correctly be mixed
with native atomics or across compiler families.  This should allow
for inline (NATIVE or SYNC) atomics with far more unknown compilers
than a more restrictive rule (such as: SYNC is compatible only with
SYNC from the same compiler family, and not with NATIVE).  This edge
case should never arise for same-family CXX and the current logic is
unchanged for the case of a mixed-family CC/CXX pair in which one or
both normally use 64-bit __sync atomics.

In addition to the preprocessor logic, this commit adds more
documentation to the "Atomics HOW TO" and makes minor edits to the
existing documentation.