fixup: atomics: use GCC asm with modern Sun com...
This commit adds a patch hunk accidentally omitted when support was
added for use of GCC asm on modern Sun compilers.

The omission was harmless except that it would have caused link
failures if one built GASNet with a "modern" Sun C and then built a
client with an "ancient" Sun C/C++.  In that case the client would
have erroneously assumed the presence of a compatible "special"
atomics implementation in the library.