trace: trace TM in RADDR{FMT,STR}()
This commit adds tracing of the "tm" argument to Put, Get and Ratomic
operations which use the GASNETI_RADDR{FMT,STR}() macros.  Where those
would trace a (node,addr) pair, we now have a (tm:rank, addr) triple.

Currently, the 'tm' is traced by a raw pointer.  However, that is
isolated to a single macros which should be updated to print an
appropriate GUID when one exists.  Similarly the "tm:rank" notation
is factored in case we want to replace it.