bug3750: inline the fast-path of dynamic thread lookup
gasnete_mythread() is used by to perform dynamic thread lookup in PAR
mode for many GASNet operations. On non-threadinfo-opt platforms
(ie most modern platforms), this is used directly for thread identification,
in lieu of the explicit thread-info passing goop (GASNET_BEGIN_FUNCTION
and friends) which compile away to nothing there.

Until now this lookup operation generally incurred a "real" function call
overhead (in the absence of LTO or other aggressive multi-module compiler
optimizations). This change moves the fast-path of this query (where the
thread has previously been seen by GASNet) into headers, allowing access to
the thread-specific register to be directly inlined into the caller.

On x86_64/Linux (dirac) with gcc-7.2.0 this was observed to speed this
query operation by 0.4ns, a 14% improvement.