atomics: fix corner case in atomics compatibility
This commit repairs a minor flaw in the preprocessor logic for
compatibility of atomics between an unknown compiler and the one used
to build the GASNet libraries.  The problem could arise only on ILP32
platforms with no native 64-bit atomics, such as 32-bit ARM.

The solution is to remove a case in the #if-ladder which assumes
equality between the implementations used by build and current
compilers is a NO-OP.  With the removal of this case, GENERIC and OS
atomics are "WANT"ed explicitly rather than implicitly, resolving the
problem.  It is worth noting that with the recent addition of explicit
handling of SPECIAL atomics, the removed case had become redundant
(with all possible cases values having their own rungs in the ladder).