fixup: atomics: fix corner case in atomics comp...
This commit corrects a mistake made in the previous commit to this
file.  Specifically, the removal of the equality case *did* change the
logic in unintended ways.  So the "lost" checks are now performed
explicitly in the SYNC and NATIVE cases, which have been merged (since
their identical bodies had become non-trivial).