coll: remove cloned "default" entry points
This commit removes `#if 0`-guarded instances of the default
collective entry points.  These appear to have been preserved as a
fall-back if the autotuner logic was removed.  However, the
corresponding algorithm selection is also preserved *within* the
autotuner as the defaults.

Attempts to enable and compile these clones was successfully, but the
resulting builds hang in testcoll/testcollperf, indicating that
something (not worth tracking down) had fallen into disrepair.