improve gasnete_event_check()
This commit improves the gasnete_event_check() macro.  The logic now
has distinct failure messages for the cases of free EOP, free IOP, and
for anything that does not appear to be a valid root event.

Previously an event was considered to be either a live EOP or IOP, and
when not a live EOP then it was treated as a live IOP.  The case of a
*free* EOP would lead to failure in gasneti_memcheck(), which would
suggest to a developer a memory corruption or invalid pointer, when
the actual issue was a use-after-sync.