Ratomic: non-fetching CAS
This commit introduces `GEX_OP_CAS`, which is a non-fetching
compare-and-swap operation, and renames `GEX_OP_CSWAP` to
`GEX_OP_FCAS` for consistency (retaining the former as a deprecated
alias).

In the case of aries-conduit, the NIC provides non-fetching CAS that
requires twice the normal alignment.  Therefore, the implementation
uses the NIC's fetching CAS and discards the result without additional
branches.