aries/gemini: fix under-allocation of AM buffers
This commit ensures that reply buffers are sized appropriately for the
maximum sized payload that may be allocated to an NP-AM Request.  This
is necessary because the reply buffers are also used to construct the
outgoing request.

The issue being corrected is that the logic for the slot-based
allocator operates in units that are always a power-of-two.  For
non-default configuration this means that the largest block the
allocator may grant for NP-AM is marginally larger than the
configure-requested MaxMedium plus header.  For instance, in a build
configured using --with-aries-max-medium=64896, the allocator gives
out maximum 65536-byte buffers and NPAM is advertising that size,
minus the header space.

Note that the marginal increase in memory use is in an "O(1)"
allocation, not something that scales with process count.  However,
since the additional space is NOT available to the user other than via
NPAM Medium Requests, the README is updated to encourage use of "good"
values.