Update gasneti_getSegmentInfo() alignment check
This commit makes a small change to gasneti_getSegmentInfo() to enable
a sanity check for segment alignment only if the configured using
--enabled-aligned-segments *and* the client used gasnet_attach() to
allocate the segment.  This aligns the implementation with the
decision to provide segment alignment only to legacy clients, while
retaining the sanity check when it is applicable.

This change resolves numerous failures seen in testing with UPCR-3 on
FreeBSD, where our testers --enable-aligned-segments.