aries/gemini: create initial TM early in init
This commit modifies the logic in aries/gemini-conduit to ensure that
the very first TM is created before any AM-based bootstrap collectives
run, and that it is used explicitly (rather than via gasneti_THUNK_TM)
for those collectives and for exit-time AMs.

This resolves a recent regression in which a SEGV would occur at
startup due to NULL gasneti_THUNK_TM in a in-NBRHD query.

Additionally this commit removes a work-around for NULL
gasneti_THUNK_TM in gasnetc_Token_Info() and special cases for NULL
'tm' arguments to Short and Medium Requests.