TM-ification of Core APIs
This commit implements translation of (tm,rank) pairs to jobrank in all
AM injection functions (including Prepare), and in Segment_QueryBound.

As part of this work, the nbrhd/pshm functions have been modified to
all take a jobrank instead of (tm,rank).  This is based on the
observation that their callers must always generate a jobrank as a
prerequisite to deciding to call them (except smp-conduit which calls
unconditionally).

An effort has been made to use 'rank' and 'jobrank' uniformly,
replacing 'dest' in many instances.  However, it is unlikely that they
have all been updated.