VIS: fix degen ops with GASNETE_PUTGET_ALWAYSREMOTE
UPCR defines GASNETE_PUTGET_ALWAYSREMOTE and this was breaking
UPCR-EX VIS for trivially degenerate cases handled in the public header.

Factor the public header handling of trivially degenerate cases
further and add a special case to deal with GASNETE_PUTGET_ALWAYSREMOTE.
The latter should be removed if/when we remove that brittle option.