coll: repair thread safety in collective progress
With the introduction of collectives progress in AMPoll(), an issue
with thread safety of the collectives progress engine has come to
light, and is addressed in this commit.

The issue is that the one-thread-polls logic meant to reduce lock
contention makes the assumption that only a single thread with
"my_local_image" set to 0 will enter the progress engine.

While it is correct that only one thread with my_local_image=0 may
initiate collectives, it is possible that many threads (including the
IBV progress thread, and Legion's progress thread), may have that same
value.  This is because 0 is the default prior to gasnet_coll_init().
Now that collectives progress is permitted from an AMPoll(), these
threads may progress concurrently and true locking is required.

Note that this resolution is a partial step towards restoring the
original "ALL_THREADS_POLL" behavior, which is currently disabled (and
known to have suffered some bitrot).  This will be necessary not only
because the concept of an "image" number will not carry forward into
GASNet-EX, but also because it is fundamentally broken for the case of
a team not containing a process's first image!

Additionally, this commit adds an explanatory comment for some
non-obvious thread safety, which was initially suspected as the cause
of duplicate frees which are resolved by this commit.