testcoll: add polling threads option to PAR case
This commit adds a '-p N' optional argument to testcoll, modeled after
that in, for instance, testam.  A significant fraction of this commit
deals with making the allowance for optional arguments, and to
updating the usage info to omit PAR-only usage in a non-PAR build.

Additionally, this commit adds a run with the new option to the
harness.conf.

(cherry picked from commit d9a9d8c718e5509fe1d65a659eb31c2f42bea5d5)

Conflicts:
	tests/testcoll.c