coll: repair split's 'relrank' semantics
This commit repairs the implementation of gasnet_coll_team_plot() to
match its specification with respect to 'relrank'.  Specifically this
commit adds the required sorting behavior which was missing from the
prior implementation (resulting in taking relrank as the absolute rank
in the new team, even if this left gaps or duplicates among the ranks).

While gasnet_teams.tex leaves the case of duplicate 'relrank'
undefined, this implementation yields the same behavior as MPI (and
upcoming GEX) specifications of "Split".  Specifically, ties among
'relrank' (a.k.a. 'key' in MPI) are broken using the ranks in the
parent team.

Additionally, a couple of local variables of type uint32_t are changed
to gex_Rank_t since they are used to count or index over team members.

Finally, testteam.c is updated to trigger the "gaps" behaviour which
caused errors in the prior implementation.