trace: build gasneti_dynsprintf() unconditionally
This commit rearranges placement of gasneti_dynsprintf() and
associated support to allow it to build unconditionally, where
previously it was only built when trace and/or stats were enabled.
This is a prerequisite to use of gasneti_dynsprintf() for formating of
TM GUIDs, which takes place in fatalerror messages (not just traces).

This commit makes no changes to the code, just moving it outside of
existing "#if GASNETI_STATS_OR_TRACE" blocks and adding an associated
comment.