coll: address thread safety issues (bit rot)
This commit works around the "bit rot" in the ALL_THREADS_POLL
support, in which multi-address Exchange and GatherAll will hang
in PAR builds for NO/NO sync flags.  The multi-address collectives and
NO/NO sync support will both be removed in GEX, leaving no motivation
to track down and fix the root problem (I have tried but failed).  So,
this is a temporary measure to allow testing until use of these
operations is eliminated and this support is removed.

Additionally, several lock-held assertions are added, with one lock
moved from function scope to file scope to enable this.