coll: revise locking in poll_fn for weak memory...
This commit makes a small change to the (unlocked) active list
traversal by the collectives progress function to allow for platforms
with weak memory ordering.  Specifically the traversal now ends at the
"original" list end, rather than potentially visiting operations
appended to the list concurrently with the traversal.  This is
necessary for correctness in a weak memory order scenario because the
pointer to any new op could be observed out-of-order with respect to
the data it references.

The alternative solution would be to add additional memory fences,
which does not seem worth to cost just to allow a concurrently created
operation to be polled one iteration sooner than would occur anyway.