coll: revise collectives initialization
This commit splits `gasnete_coll_init()` into two portions.

The name `gasnets_coll_init()` is retained as the implementation of
the client-facing `gasnet_coll_init()` interface, which is retained to
support registration of reduction functions and a small amount of
image-related initialization.

The gex-relevant portions of the teams/collectives initialization
moves to a new function, `gasnete_coll_init_subsystem()`, which is
now called from `gasnete_init()`.

Additionally, testsplit.c is updated to no longer call
`gasnet_coll_init()` (which was needed to permit calls to Split) and
the corresponding include of `gasnet_coll.h` is removed.