coll-ex: support for multi-field reduction
This commit generalizes (with surprisingly little effort) the initial
algorithm for reduction to support multi-field (dt_cnt > 1).

The presence of multi-field support has the effect of merging the two
documented limitations (on dt_cnt and dt_sz) into a single limitation
on their product.  The specification is updated accordingly.

To improve test coverage, testcollperf is changed by this commit to
perform a computation duplicating that of the reference collectives,
in order to test as large a dt_cnt as possible.  This is a temporary
measure until the artificial limit is removed.