coll-ex: GEX-ification update to Reduce
This commit follows-up on the implementation of
gex_Coll_ReduceToOneNB() to remove several "legacy" aspects.  In
particular this commit introduces (tm,rank) based versions of a couple
of key "p2p" interfaces (to permit removal of "REL2ACT" and indirect
use of THUNK_TM) and conversion of the binomial tree helpers from
'team' to 'tm'.

In other words, this updates the one Reduce algorithm to look the way
I want future GEX collective algorithms to look.