Rename remaining gasnete_ thread data to gasneti_
This commit renames several `gasnete_`/`GASNETE_` identifiers related
to thread-specific data to the `gasneti_`/`GASNETI_` namespace.  This
includes the type `gasneti_threaddata_t` and the family of macros used
for the THREADINFO_OPT "goop".  Additionally, `gasnete_mythread()` is
renamed to `_gasneti_mythread_slow()` to emphasize that it is for
internal use when no other source of the information is available.