Legion CI: allow removing reliance on TOP_BUILDDIR
This commit makes changes to the external-legion harness suite to
permit correct operation when `$(TOP_BUILDDIR)` is available at harness
compile time but NOT when the built executables run.

The sole dependence on `$(TOP_BUILDDIR)` was through the placement of
the GASNet install (or symlink to it) within `$(TOP_BUILDDIR)/legion`.
Therefore, the main change in this commit is to relocate the default
location to `$(HARNESS_WORKDIR)/legion`.  Note that for the case that
no `-workdir` argument is passed to harness, this means NO CHANGE to
the location for a new installation.

In the case that `$(LEGION_GASNET)` names an existing GASNet
installation, the Makefile has been reworked to use that variable
(stripped of leading and trailing whitespace) directly rather than
creating a symlink.  This is the only non-trivial externally-visible
change in behavior for the case that no `-workdir` argument is passed.
However, it does account for the bulk of the changed lines.

(cherry picked from commit 9a46ca3d57840a880813dfa6a84b8bba52af813d)