Chapel CI: allow removing reliance on TOP_BUILDDIR
This commit modifies the build logic for the external-chapel CI suite
to perform a `make install` in the case that `$(HARNESS_WORKDIR)` is
not the same as `$(TOP_BUILDDIR)`.  Since those are the same by
default, there is no change in behavior unless an explicit `-workdir`
argument is passed to harness.

If one *does* pass an explicit `-workdir` argument to harness, the
Chapel compiler and runtime are now installed within the workdir,
allowing for the possibility that the builddir is removed between
the time executables are built and when they are run.

(cherry picked from commit e51460d3dc8157f7e23e246dea15d6f61b065ec9)