Remove the deprecated shmem-conduit
Despite its misleading name, shmem-conduit relied critically upon
the hardware DSM support of the obsolete Cray X1 and SGI Altix NUMAlink
networks, because it required shmem_ptr() to always return valid load/store-
addressable pointers to remote memory.

The systems in question are all obsolete and retired, and due to lack of
hardware access the conduit can no longer be maintained. It was officially
deprecated in Nov 2016, and now is removed.