coll-ex: "gexify" several p2p APIs
This commit updates several collectives point-to-point utility
functions to take `(tm,rank)` instead of a jobrank, to accept `lc_opt`
(where applicable) and `flags` arguments, and to return an `int` (for
IMMEDIATE failure).  Additionally, the `src` arguments are given a
`const` qualifer.

Functions which resolve to a single AM injection call have been
hoisted to the header and made INLINE.  This has necessitated moving
the definitions of handler indices between headers.

The GASNETI_THREAD_FARG has been propogated through the updated
functions as well, avoiding a dynamic thread lookup in the AM
injection.

The "eager" implementation of `gex_Coll_ReduceToOneNB()` is updated
for the new eager put signature, including use of IMMEDIATE to avoid
stalling in the presence of back pressure.