Legion CI: fix launcher-symlink regression
This commit restores the previous placement of double-quotes which
were moved (without good cause) in the previous commit to this file.
The result of the previous move was that the trailing space present in
`$(LEGION_CONDUIT)` was becoming part of the target of the `launcher`
symlink when the conduit was neither smp nor udp.