coll: remove some unused geom fields
This commit removes several unused fields from the "local view"
geometry struct.  In particular info regarding grandchildren, and DFS
traversal are removed.  The former is not used (nor is any likely use
envisioned) and the later is no longer relevant because all trees are
constructed with in-order DFS sequence (for either forward or reverse
visitation of children).