Default to GASNET_NODEMAP_EXACT=1
This commit changes the default value for the environment variable
`GASNET_NODEMAP_EXACT` from 0 to 1.  This change may result in marginally
slow startup, but provides the highest quality supernode assignments.

Some additional info appears in Bug 3770, an RFE that suggests how
one may be able to restore the old default and detect when the
generated map is "bad".