Exclude auxsegsz from mmapLimit
In EX, the auxseg is a separate mmap from the client segment, so
it's size should not be subtracted from the mmapLimit (which in
particular breaks the new GASNET_MAX_SEGSIZE=/p guarantee of providing
at least that much MaxLocal(), subject only to resource constraints).

Also, ensure that when GASNET_MAX_SEGSIZE is set very small that
gasneti_mmapLimit() still probes to return a size large enough to
accomodate the auxsegment.