Chapel CI: fix regression w/ non-bash shells
This commit repairs a regression introduced with the recent changes to
install Chapel to the workdir (when distinct from the blddir).  That
change caused a problem for shells other than bash and zsh, because
only those two shells can source `util/setchplenv.*` from a cwd other
than `$CHPL_HOME`.  Therefore this commit changes the test-building
makefile recipe to change directory, source the file, and then change
back to the original directory.  That last step us done using gmake's
`$(CURDIR)`, which holds the directory in which make is running (after
processing of `-C` options).