NP-AM: POST(sd->_thread) in Nbrhd Prepare
This commit makes use of the `_thread` field always present in the
SrcDesc structure to eliminate use of THREAD_FARG from the NP-AP
implementation used for in-nbrhd (both PSHM and true loopback).

This saves one function call argument on ABIs "weak" enough that we
want the THREADINFO_OPT, as well generally simplifying the code.