NP-AM: no FARG in gasnetc_AM_PrepareReply*()
This commit removes GASNETI_THREAD_FARG from the PrepareReply
functions, and the corresponding GET from the wrapper macros.

Except for the unlikely case that a client has GASNET_BEGIN_FUNCTION()
in their AM request handler, this change simply defers a dynamic
thread lookup to occur within the Prepare, rather than in the caller.
However, in the future this dynamic lookup can (should) be replaced
with a GASNET_THREAD_POST() using a threadinfo to be added to the
token.