Nbrhd AM: add GASNETI_THREAD_FARG
This commit adds GASNETI_THREAD_FARG to some internal functions within
the Nbrhd (both PSHM and true loopback) code for AM Request and Reply.

In the case of Requests, this eliminates a dynamic thread lookup.
In the case of Replies, this moves the dynamic thread lookup earlier
in the call chain, but gives the opportunity to later replace the
dynamic lookup with a POST of threadinfo to be stored in the token.