Remove unused THREAD_FARG handling in Reply paths
This commit removes most of the GASNETI_THREAD_FARG handling from AM
Reply paths.  Specifiably, conduits not making any use of threadinfo
in Replies no longer pass it around.  The exceptions are ibv-conduit
which does make use of TLS for lock-free resource pools, and the
gasnetc_AM{Medium,Long}V() functions.  The later are functions are
conduit-independent and thus retain a THREAD_FARG because ibv-conduit
uses it (for now).

In the future it is expected that when THREADINFO_OPT is enabled,
Reply logic will have access to a threadinfo stored in the token.
Once that is available, it will displace THREAD_FARG passing in
ibv-conduit, and the gasnetc_AM{Medium,Long}V() functions will no
longer need THREAD_FARG either.