ibv: stash threadinfo in AM token
This commit updates ibv-conduit's AM receive logic, conditional on
GASNETI_THREADINFO_OPT, to stash the threadinfo pointer in the AM
token.  This allows POSTing of this pointer in AM Reply paths,
instead of a dynamic thread lookup.

This commit also eliminates the last consumer of the THREAD_FARG
passed to AMReply{Medium,Long}V().