NP-AM: remove FARG from gasnetc_AMReply...V()
This commit removes the GASNETI_THREAD_FARG passed to
gasnetc_AMReply{Medium,Long}V(), following-up on the elimination of
ibv-conduit as its sole consumer.

This introduces a GASNETI_MYTHREAD_GET_OR_LOOKUP in gemini-conduit
(ONLY for the case of multi-domain Replies), but only to offset
elimination of a GASNETI_THREAD_GET higher in the call stack.