gemini/aries: stash threadinfo in AM token
This commit saves a copy (conditional on GASNETI_THREADINFO_OPT) in
the AM token.  POSTing of this pointer replaces dynamic thread lookups
in AM Reply paths.

The majority of the lines changed in this commit are devoted to
threading the THREAD_FARG from gasnet_AMPoll() down to the point at
which the token is finally created.  This is complicated slightly in
the presence of multi-domain by the need for some polling to take
place with domain index 0, even when that is not the calling threads's
actual domain.