aries/gemini: implement rendezvous-based AMs
This commit implements an optional rendezvous-based algorithm for AM
Requests, which increases latency to save on registered memory.  The
approach is to send a tiny amount of data to the target, used to
initiate a Get of the AM header and payload, rather than using a Put.
This allows all incoming Request to be processed with O(1) memory,
rather than the O(procs) memory used in the pre-existing algorithm
(still present, and now known as the "Eager" algorithm).

Use of the Eager vs Rendezvous algorithm is controlled by a new
GASNET_GNI_AM_RVOUS_CUTOVER environment variable and the number of
max-sized buffers used.  These are documented in the README.

Additional space (and latency) savings is possible by eliminating the
use of the notify ring in favor of a CqWrite.  However, that is future
work.